Exception handling

+++ index.js
@@ -74,6 +74,20 @@ const fileExt = (url) => {
 	return result
 };
 
+const downloadFile = async (link, fileName) => {
+	let result = ''
+	try {
+		const { stdout, stderr } = await exec(`wget -O ./uploads/${fileName} ${link}`);
+	} catch (error) {
+		console.log(String(error))
+		if (String(error).includes('ERROR 404'))
+			result = 'File not found!'
+		else
+			result = String(error)
+	}
+	return result
+}
+
 // http://localhost:8021/ocr-pdf?rotate=1&background=1&deskew=1&clean=1&url=https://transfer.sh/DFGmlvglle/Public%20WaterMassMailing-1.pdf
 app.get("/ocr-pdf", async (req, res) => {
 	if (!fs.existsSync("./uploads")) {
@@ -98,9 +112,6 @@ app.get("/ocr-pdf", async (req, res) => {
 		try {
 			if (fileExtension === 'pdf') {
 				const outputFile = `ocr-${new Date().getTime()}-${uuidv4()}.pdf`
-				const { stdout, stderr } = await exec(`wget -O ./uploads/${outputFile} ${link}`);
-				/* console.log('stdout:', stdout);
-				console.log('stderr:', stderr); */
 
 				await exec(`ocrmypdf ${params.join(' ')} --skip-text --output-type pdfa --sidecar './uploads/${outputFile.replace(/\.pdf$/gi, '.txt')}' './uploads/${outputFile}' './uploads/${outputFile}'`);
 
@@ -110,15 +121,16 @@ app.get("/ocr-pdf", async (req, res) => {
 				let imgFile = `${fileWithoutExt}.${fileExtension}`
 				const outputFile = `${fileWithoutExt}.pdf`
 				const outputTxtFile = `${fileWithoutExt}.txt`
-				const { stdout, stderr } = await exec(`wget -O ./uploads/${imgFile} ${link}`);
-				/* console.log('stdout:', stdout);
-				console.log('stderr:', stderr); */
+				let isDownloaded = await downloadFile(link, imgFile)
+				if (isDownloaded === '') {
+					if (fileExtension === 'png') sharp(`./uploads/${imgFile}`).flatten({ background: '#fff' })
 
-				if (fileExtension === 'png') sharp(`./uploads/${imgFile}`).flatten({ background: '#fff' })
+					await exec(`ocrmypdf --image-dpi 300 ${params.join(' ')} --skip-text --output-type pdfa --sidecar './uploads/${outputTxtFile}' './uploads/${imgFile}' './uploads/${outputFile}'`);
 
-				await exec(`ocrmypdf --image-dpi 300 ${params.join(' ')} --skip-text --output-type pdfa --sidecar './uploads/${outputTxtFile}' './uploads/${imgFile}' './uploads/${outputFile}'`);
-
-				res.status(200).json({ isSuccessful: true, file: outputFile })
+					res.status(200).json({ isSuccessful: true, file: outputFile })
+				} else {
+					res.status(200).json({ isSuccessful: false, file: '' })
+				}
 			}
 		} catch (error) {
 			console.log(error)
@@ -148,24 +160,27 @@ app.get("/compress-pdf", async (req, res) => {
 		try {
 			const inputFile = `ocr-${new Date().getTime()}-${uuidv4()}.pdf`
 			const outputFile = `ocr-${new Date().getTime()}-${uuidv4()}.pdf`
-			const { stdout, stderr } = await exec(`wget -O ./uploads/${inputFile} ${link}`);
-			/* console.log('stdout:', stdout);
-			console.log('stderr:', stderr); */
+			let isDownloaded = await downloadFile(link, inputFile)
 
-			await exec(`gs -sDEVICE=pdfwrite -dCompatibilityLevel=1.4 -dPDFSETTINGS=/${compressMode} -dNOPAUSE -dQUIET -dBATCH -sOutputFile='./uploads/${outputFile}' './uploads/${inputFile}'`);
+			if (isDownloaded === '') {
 
-			const inputFileSize = fs.statSync(`./uploads/${inputFile}`).size
-			const outputFileSize = fs.statSync(`./uploads/${outputFile}`).size
+				await exec(`gs -sDEVICE=pdfwrite -dCompatibilityLevel=1.4 -dPDFSETTINGS=/${compressMode} -dNOPAUSE -dQUIET -dBATCH -sOutputFile='./uploads/${outputFile}' './uploads/${inputFile}'`);
 
-			fs.unlinkSync(`./uploads/${inputFile}`)
+				const inputFileSize = fs.statSync(`./uploads/${inputFile}`).size
+				const outputFileSize = fs.statSync(`./uploads/${outputFile}`).size
 
-			res.status(200).json({
-				isSuccessful: true, file: {
-					name: outputFile,
-					originalSize: inputFileSize,
-					compressedSize: outputFileSize,
-				}
-			})
+				fs.unlinkSync(`./uploads/${inputFile}`)
+
+				res.status(200).json({
+					isSuccessful: true, file: {
+						name: outputFile,
+						originalSize: inputFileSize,
+						compressedSize: outputFileSize,
+					}
+				})
+			} else {
+				res.status(200).json({ isSuccessful: false, file: {} })
+			}
 		} catch (error) {
 			console.log(error)
 			res.status(200).json({ isSuccessful: false, file: {} })
@@ -175,3 +190,8 @@ app.get("/compress-pdf", async (req, res) => {
 		res.status(200).json({ isSuccessful: false, file: {} })
 	}
 });
+
+app.get("/", async (req, res) => {
+	await downloadFile('http://10.39.1.65:8005/dms/Ashis/attendance_f412.gif', 'attendance_f4.gif')
+	res.send('WELCOME')
+})
